refactor(api): extrai validators web-cookie + Meta AI de validation.ts
Decomposição god-file (Tier 1 / #3501), 2ª fatia. Move os 13 validators de
provider web-cookie + o builder de request Meta AI para módulos co-localizados,
byte-idênticos ao inline original (funções top-level, sem captura de estado do
dispatcher).

- providers/validation/metaAi.ts        — base62/ID-gen + buildMetaAiValidationBody (sem imports; usa globais)
- providers/validation/webProvidersA.ts — deepseek/qwen/grok/chatgpt/perplexity/blackbox
- providers/validation/webProvidersB.ts — muse-spark/adapta/claude/gemini/copilot/t3/jules/inner-ai

validation.ts: 4234 → 3027 (−1207). webProvidersB importa metaAi; host importa
os 13 validators e os injeta no SPECIALTY_VALIDATORS. Imports órfãos removidos
do host: webCookieAuth (5 símbolos) + buildJulesApiUrl.

Validação: typecheck:core, eslint (0/0), check:cycles, check:file-size (módulos
<800), 149 testes de validação existentes + 4 de caracterização — todos verdes.
